fix(tk:shared): check element exists before profile metadata extraction
